+++ dcmimage/libsrc/diqtpbox.cc
@@ -21,10 +21,10 @@
  *
  *  Purpose: class DcmQuantPixelBoxArray
  *
- *  Last Update:      $Author: meichel $
- *  Update Date:      $Date: 2002-01-25 13:32:11 $
+ *  Last Update:      $Author: joergr $
+ *  Update Date:      $Date: 2002-07-18 12:19:50 $
  *  Source File:      $Source: /export/gitmirror/dcmtk-git/../dcmtk-cvs/dcmtk/dcmimage/libsrc/diqtpbox.cc,v $
- *  CVS/RCS Revision: $Revision: 1.1 $
+ *  CVS/RCS Revision: $Revision: 1.2 $
  *  Status:           $State: Exp $
  *
  *  CVS/RCS Log at end of file
@@ -35,9 +35,9 @@
 #include "osconfig.h"
 #include "diqtpbox.h"   /* for DcmQuantPixelBoxArray */
 
-static int sumcompare(const void *x1, const void *x2) 
+static int sumcompare(const void *x1, const void *x2)
 {
-  return (*(DcmQuantPixelBoxPointer *) x2)->sum 
+  return (*(DcmQuantPixelBoxPointer *) x2)->sum
        - (*(DcmQuantPixelBoxPointer *) x1)->sum;
 }
 
@@ -46,7 +46,7 @@ void DcmQuantPixelBoxArray::sort(unsigned long boxes)
 #ifdef DEBUG
   assert(boxes <= length);
 #endif
-  qsort((char *)array, boxes,  sizeof(DcmQuantPixelBoxPointer), sumcompare);
+  qsort((char *)array, (unsigned int)boxes,  sizeof(DcmQuantPixelBoxPointer), sumcompare);
 }
 
 DcmQuantPixelBoxArray::DcmQuantPixelBoxArray(unsigned long entries)
@@ -67,7 +67,10 @@ DcmQuantPixelBoxArray::~DcmQuantPixelBoxArray()
  *
  * CVS/RCS Log:
  * $Log: diqtpbox.cc,v $
- * Revision 1.1  2002-01-25 13:32:11  meichel
+ * Revision 1.2  2002-07-18 12:19:50  joergr
+ * Added explicit type casts to keep Sun CC 2.0.1 quiet.
+ *
+ * Revision 1.1  2002/01/25 13:32:11  meichel
  * Initial release of new color quantization classes and
  *   the dcmquant tool in module dcmimage.
  *
